Include `LintContext` tests in interface tests
