+++ test/runtests.jl
@@ -1,11 +1,10 @@
 using ReLint
+using Argus: Rule, @pattern
 using Test
 
 include(joinpath(@__DIR__, "common.jl"))
 
+include(joinpath(@__DIR__, "interface.jl"))
 include(joinpath(@__DIR__, "fatal.jl"))
 include(joinpath(@__DIR__, "violations.jl"))
 include(joinpath(@__DIR__, "recommendations.jl"))
-
-# include(joinpath(@__DIR__, "lint_context_tests.jl"))
-include(joinpath(@__DIR__, "interface.jl"))
